basic function conplete
