+++ .gitignore
@@ -1,3 +1,4 @@
+radiolabdocker.egg-info
 radiolabdocker/__pycache__/*
 /docker-compose.yml
 /build
