+++ bin/radiolabdocker
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 import os.path as op
 from os import makedirs
 from radiolabdocker.BuildImage import buildIMAGE, buildSeq
