+++ config/bash_config/bashrc
@@ -142,8 +142,3 @@ if ! shopt -oq posix; then
     . /etc/bash_completion
   fi
 fi
-
-# Set DISPLAY
-if [ -z ${DISPLAY} ]; then
-    export DISPLAY=:0
-fi
