+++ radiolabdocker/CheckStat.py
@@ -1,18 +1,13 @@
-import docker
-
-base = 'radiolab_xpra'
-tag = 'latest'
-path = './build'
-
 def checkContainerStat(check_container: str):
     """
     Return the existance and running status of a given container name.
     :param: check_container: the name of a contariner.
     :return: (existance (bool), running status (Running, Exit, None if not exist))
     """
+    import docker
     client = docker.from_env()
-    container_ls = client.containers.list()
-    container_stat = {container.name: container.status for container in container_ls}
+    container_ls = client.containers.list(all = True)
+    container_stat = {container.name: [container.status, container.id] for container in container_ls}
     contariner_name = container_stat.keys()
     # Check existance and status
     if check_container in contariner_name:
@@ -20,7 +15,7 @@ def checkContainerStat(check_container: str):
         stat = container_stat[check_container]
     else:
         exist = False
-        stat = None
+        stat = None, None
     #
     return(exist, stat)
 
@@ -30,6 +25,7 @@ def checkImageStat(check_img: str):
     :param: check_image: the name of a image.
     :return: (existance (bool), tags (dict))
     """
+    import docker
     client = docker.from_env()
     image_ls = client.images.list()
     img_tags = { img.tags[0].split(':')[0]:[ tag.split(':')[1] for tag in img.tags ] for img in image_ls }
